Fix MaintenanceOrderDetail.tsx wrapper and complete maintenance-orders feature migration
